# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Base Plate API",
        "description": "A product of Appetiser",
        "contact": {
            "name": "API Support",
            "email": "dennis.agulo@appetiser.com.au"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://128.199.193.26/api/v1",
            "description": "Base Endpoint"
        }
    ],
    "paths": {
...
        "/password/email": {
            "post": {
                "tags": [
                    "Password"
                ],
                "summary": "Send reset password email with token",
                "operationId": "App\\Http\\Controllers\\API\\V1\\ApiPasswordResetsController::sendResetCodeEmail",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email Address",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Invalid Input"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "200": {
                        "description": "Request OK"
                    }
                },
                "security": []
            }
        },
        "/password/reset": {
            "post": {
                "tags": [
                    "Password"
                ],
                "summary": "Reset user's password",
                "operationId": "App\\Http\\Controllers\\API\\V1\\ApiPasswordResetsController::resetPassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email Address",
                                        "type": "string"
                                    },
                                    "token": {
                                        "description": "Reset Token",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "description": "Confirm Password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Invalid Input"
                    },
                    "404": {
                        "description": "Reset Token Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "200": {
                        "description": "Request OK"
                    }
                },
                "security": []
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}